# specs/001-mvp-closed-loop/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PPTagent MVP API
  version: 0.1.0-mvp
  description: |
    REST API for PPTagent MVP 业务闭环（生成 / 知识库 / Agent 进化）。
    与 WebSocket 事件契约（events.yaml）配合使用，错误码遵循 RFC 7807。
servers:
  - url: http://localhost:8000/api/v1
    description: Local dev
  - url: https://api.pptagent.example.com/api/v1
    description: Production (placeholder)
security:
  - ApiKeyAuth: []
paths:
...
  /materials:
    get:
      summary: 素材库检索（FR-031：BM25 + 嵌入向量双路召回）
      description: |
        `query` 可选；不传则按筛选条件返回最近被复用的素材。
        `visual_types` / `industry_tags` / `source_sample_ids` 三个筛选维度可任意组合。
        结果按"BM25 × 0.4 + 视觉类型 × 0.2 + 嵌入余弦 × 0.4"加权相关度排序。
      operationId: searchMaterials
      tags: [Materials]
      parameters:
        - in: query
          name: query
          schema: { type: string, maxLength: 200 }
        - in: query
          name: visual_type
          schema:
            type: string
            enum: [cover, toc, architecture, flowchart, data, body, closing, mixed]
        - in: query
          name: industry
          schema: { type: string, description: '行业标签，可重复' }
        - in: query
          name: source_sample_id
          schema: { type: string, format: uuid }
        - in: query
          name: include_orphan
          schema: { type: boolean, default: false }
        - in: query
          name: page
          schema: { type: integer, default: 1, minimum: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MaterialSearchResult' }
  /materials/{asset_id}:
    parameters:
      - in: path
        name: asset_id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: 素材详情（含 SVG 原文 / 结构 / 风格指纹）
      operationId: getMaterial
      tags: [Materials]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SlideAssetDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: 删除素材（FR-037 不级联 sample）
      operationId: deleteMaterial
      tags: [Materials]
      responses:
        '204': { description: OK }
        '409':
          description: 素材正被某草稿引用
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /materials/{asset_id}/insert:
    parameters:
      - in: path
        name: asset_id
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: 将素材插入指定草稿（FR-033 / FR-034 风格归一开关）
      operationId: insertMaterialIntoDraft
      tags: [Materials]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InsertMaterialRequest' }
      responses:
        '201':
          description: 插入成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DraftSlide' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '423':
          description: 草稿被其他用户锁定（FR-035 单写者）
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ─── US6: 草稿 (Drafts) ───
...
  /data/export:
    post:
      summary: 导出我的全部数据（FR-018）
      operationId: exportData
      tags: [DataLifecycle]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DataExportRequest' }
      responses:
        '202':
          description: 已入队打包
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportJob' }
...
  /data/delete-all:
    post:
      summary: 删除我的全部数据（FR-019）
      operationId: deleteAllData
      tags: [DataLifecycle]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DataDeleteRequest' }
      responses:
        '202':
          description: 已入队
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string, format: uuid }
                  estimated_hard_delete_at: { type: string, format: date-time }
        '400':
          description: confirm_phrase 不匹配
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ─── Security ───
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 形如 `Bearer <key>`（最小权限 IV + 速率限制）
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: 状态冲突（如取消已完成任务）
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    ExportJob:
      type: object
      properties:
        id: { type: string, format: uuid }
        status:
          type: string
          enum: [pending, running, ready, failed]
        download_url: { type: string, format: uri, nullable: true }
        expires_at: { type: string, format: date-time, nullable: true }
...
    Problem:
      type: object
      description: RFC 7807
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        code: { type: string, description: PPTagent 自定义错误码，详见 error-codes.yaml }
        request_id: { type: string, format: uuid, description: 关联 X-Request-Id }
        trace_id: { type: string, description: OTel trace ID }
        errors:
          type: array
          description: 字段级错误（422 验证失败时填充）
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }

    # ─── Pagination ───
...
    DataExportRequest:
      type: object
      required: [confirm_phrase, acknowledged_at]
      properties:
        include_raw:
          type: boolean
          default: true
          description: 包含原始文件
        include_parse:
          type: boolean
          default: true
          description: 包含解析结果
        include_preferences:
          type: boolean
          default: true
          description: 包含偏好规则
        include_traces:
          type: boolean
          default: false
          description: 包含生成轨迹
        confirm_phrase:
          type: string
          enum: [EXPORT ALL MY DATA]
          description: 二次确认；必须字面量匹配
        acknowledged_at:
          type: string
          format: date-time
    DataDeleteRequest:
      type: object
      required: [confirm_phrase, acknowledged_at]
      properties:
        confirm_phrase:
          type: string
          enum: [DELETE ALL MY DATA]
          description: 二次确认；必须字面量匹配
        acknowledged_at:
          type: string
          format: date-time

    # ─── API Key DTO ───
...
    SlideVisualType:
      type: string
      enum: [cover, toc, architecture, flowchart, data, body, closing, mixed]
...
    DraftSlideSourceType:
      type: string
      enum: [reused, generated, manual]
    SlideAsset:
      type: object
      description: FR-030 抽取出的可复用素材（按页为最小单位）
      properties:
        id: { type: string, format: uuid }
        owner_id: { type: string, format: uuid }
        source_sample_id: { type: string, format: uuid, nullable: true }
        source_page_index: { type: integer, minimum: 1 }
        visual_type: { $ref: '#/components/schemas/SlideVisualType' }
        content_type: { type: string, enum: [svg, text, image, chart] }
        slide_title: { type: string, nullable: true }
        text_content: { type: string, nullable: true }
        thumbnail_url: { type: string, nullable: true }
        industry_tags:
          type: array
          items: { type: string }
        is_orphan: { type: boolean }
        is_archived: { type: boolean }
        reuse_count: { type: integer }
        last_reused_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    SlideAssetDetail:
      allOf:
        - $ref: '#/components/schemas/SlideAsset'
        - type: object
          properties:
            style_features: { type: object, additionalProperties: true, nullable: true }
            svg_payload: { type: object, additionalProperties: true, nullable: true }
            structure_payload: { type: object, additionalProperties: true, nullable: true }
    MaterialSearchResult:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/SlideAsset' }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
        has_next: { type: boolean }
        next_url: { type: string, nullable: true }
    InsertMaterialRequest:
      type: object
      required: [draft_id, slide_order, enable_style_normalize]
      properties:
        draft_id: { type: string, format: uuid }
        slide_order: { type: integer, minimum: 1 }
        enable_style_normalize: { type: boolean, default: true, description: FR-034 风格归一开关 }
        expected_revision: { type: integer, minimum: 0, description: 草稿乐观锁 }

    # ─── US6 草稿 ───
...
    DraftSlide:
      type: object
      properties:
        id: { type: string, format: uuid }
        slide_order: { type: integer, minimum: 1 }
        source_type: { $ref: '#/components/schemas/DraftSlideSourceType' }
        source_asset_id: { type: string, format: uuid, nullable: true }
        source_stage_id: { type: string, format: uuid, nullable: true }
        thumbnail_url: { type: string, nullable: true }
        style_normalized: { type: boolean }
        normalized_failed: { type: boolean }
        source_attribution: { type: string }
...